items/requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
pyjwt
